backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
google-generativeai
python-dotenv
